# Fix API timing out after 15 seconds