--- a/db/migrate/20231205121051_add_user_id_created_at_index_to_hits.rb
+++ b/db/migrate/20231205121051_add_user_id_created_at_index_to_hits.rb
@@ -1,0 +1,5 @@
+class AddUserIdCreatedAtIndexToHits < ActiveRecord::Migration[7.0]
+  def change
+    add_index :hits, [:user_id, :created_at]
+  end
+end