--- a/db/migrate/20231205121414_remove_user_id_index_from_hits.rb
+++ b/db/migrate/20231205121414_remove_user_id_index_from_hits.rb
@@ -1,0 +1,5 @@
+class RemoveUserIdIndexFromHits < ActiveRecord::Migration[7.0]
+  def change
+    remove_index :hits, :user_id
+  end
+end