--- a/db/schema.rb
+++ b/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2023_12_05_120232) do
+ActiveRecord::Schema[7.0].define(version: 2023_12_05_121414) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
@@ -19,7 +19,7 @@ ActiveRecord::Schema[7.0].define(version: 2023_12_05_120232) do
     t.string "endpoint", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-    t.index ["user_id"], name: "index_hits_on_user_id"
+    t.index ["user_id", "created_at"], name: "index_hits_on_user_id_and_created_at"
   end
 
   create_table "users", force: :cascade do |t|