User#count_hits method's response time is around 500ms and is called 50 times per second (on average).

This PR fixes it by adding an index in the hits table in the user_id and created_at columns.

Old:

```
acme_api_development=# EXPLAIN ANALYZE SELECT COUNT(*) FROM "hits" WHERE "hits"."user_id" = 1 AND (created_at > '2023-12-01 00:00:00');
                                                             QUERY PLAN
------------------------------------------------------------------------------------------------------------------------------------
 Aggregate  (cost=12.65..12.66 rows=1 width=8) (actual time=0.020..0.020 rows=1 loops=1)
   ->  Bitmap Heap Scan on hits  (cost=4.18..12.65 rows=1 width=0) (actual time=0.013..0.013 rows=0 loops=1)
         Recheck Cond: (user_id = 1)
         Filter: (created_at > '2023-12-01 00:00:00'::timestamp without time zone)
         ->  Bitmap Index Scan on index_hits_on_user_id  (cost=0.00..4.18 rows=4 width=0) (actual time=0.011..0.011 rows=0 loops=1)
               Index Cond: (user_id = 1)
 Planning Time: 0.411 ms
 Execution Time: 0.091 ms
(8 rows)
```

New: 

```
acme_api_development=# EXPLAIN ANALYZE SELECT COUNT(*) FROM "hits" WHERE "hits"."user_id" = 1 AND (created_at > '2023-12-01 00:00:00');
                                                                      QUERY PLAN
------------------------------------------------------------------------------------------------------------------------------------------------------
 Aggregate  (cost=8.17..8.18 rows=1 width=8) (actual time=0.004..0.004 rows=1 loops=1)
   ->  Index Only Scan using index_hits_on_user_id_and_created_at on hits  (cost=0.15..8.17 rows=1 width=0) (actual time=0.002..0.002 rows=0 loops=1)
         Index Cond: ((user_id = 1) AND (created_at > '2023-12-01 00:00:00'::timestamp without time zone))
         Heap Fetches: 0
 Planning Time: 0.243 ms
 Execution Time: 0.019 ms
(6 rows)
```

Other future possible solutions:

1) If the previous months' data is unimportant, we can delete the old hits.
2) Cache (Example: rails counters)
3) Rate Limiting Middleware. (Ex: rack-attack)